// src/app/shared/interface/ide/code-mirror.interface.ts
export interface CodeMirrorOptions {
  addModeClass?: boolean;
  allowDropFileTypes?: any;
  autocapitalize?: boolean;
  autocorrect?: boolean;
  autofocus?: boolean;
  configureMouse?: string;
  coverGutterNextToScrollbar?: boolean;
  cursorBlinkRate?: number;
  cursorHeight?: number;
  cursorScrollMargin?: number;
  direction?: string;
  disableInput?: boolean;
  dragDrop?: boolean;
  electricChars?: boolean;
  extraKeys?: string;
  firstLineNumber?: number;
  fixedGutter?: boolean;
  flattenSpans?: boolean;
  gutters?: [string];
  historyEventDelay?: number;
  indentUnit?: number;
  indentWithTabs?: boolean;
  inputStyle?: string;
  keyMap?: string;
  lineNumbers?: boolean;
  lineSeparator?: string;
  lineWiseCopyCut?: boolean;
  lineWrapping?: boolean;
  lint?: boolean;
  maxHighlightLength?: number;
  mode?: string | object;
  moveInputWithCursor?: boolean;
  pasteLinesPerSelection?: boolean;
  phrases?: string;
  pollInterval?: number;
  readOnly?: boolean;
  resetSelectionOnContextMenu?: boolean;
  rtlMoveVisually?: boolean;
  scrollbarStyle?: string;
  selectionsMayTouch?: boolean;
  showCursorWhenSelecting?: boolean;
  singleCursorHeightPerLine?: boolean;
  smartIndent?: boolean;
  specialChars?: any;
  spellcheck?: boolean;
  tabSize?: number;
  tabindex?: any;
  theme?: string;
  undoDepth?: number;
  value?: any;
  viewportMargin?: number;
  wholeLineUpdateBefore?: boolean;
  workDelay?: number;
  workTime?: number;
}